struct and enum conversions
