+++ src/Graphics/Resources/Fence.cs
@@ -10,9 +10,9 @@ namespace MoonWorks.Graphics
 	/// The Fence object itself is basically just a wrapper for the Refresh_Fence. <br/>
 	/// The internal handle is replaced so that we can pool Fence objects to manage garbage.
 	/// </summary>
-	public class Fence : RefreshResource
+	public class Fence : SDL_GpuResource
 	{
-		protected override Action<nint, nint> QueueDestroyFunction => Refresh.Refresh_ReleaseFence;
+		protected override Action<nint, nint> ReleaseFunction => Refresh.Refresh_ReleaseFence;
 
 		internal Fence(GraphicsDevice device) : base(device)
 		{
