+++ src/Graphics/SDL_GpuResource.cs
@@ -1,17 +1,16 @@
 using System;
-using System.Runtime.InteropServices;
 using System.Threading;
 
 namespace MoonWorks.Graphics;
 
-public abstract class RefreshResource : GraphicsResource
+public abstract class SDL_GpuResource : GraphicsResource
 {
 	public IntPtr Handle { get => handle; internal set => handle = value; }
 	private IntPtr handle;
 
-	protected abstract Action<IntPtr, IntPtr> QueueDestroyFunction { get; }
+	protected abstract Action<IntPtr, IntPtr> ReleaseFunction { get; }
 
-	protected RefreshResource(GraphicsDevice device) : base(device)
+	protected SDL_GpuResource(GraphicsDevice device) : base(device)
 	{
 	}
 
@@ -19,11 +18,11 @@ public abstract class RefreshResource : GraphicsResource
 	{
 		if (!IsDisposed)
 		{
-			// Atomically call destroy function in case this is called from the finalizer thread
+			// Atomically call release function in case this is called from the finalizer thread
 			var toDispose = Interlocked.Exchange(ref handle, IntPtr.Zero);
 			if (toDispose != IntPtr.Zero)
 			{
-				QueueDestroyFunction(Device.Handle, toDispose);
+				ReleaseFunction(Device.Handle, toDispose);
 			}
 		}
 		base.Dispose(disposing);
